Add custom text when server starts
I'm using chalk to allow for easy coloration of the printed text.  This
doesn't seem to work in bash, but it does in windows consoles.

The npm convention is that because I added a "prestart" script, it will
be run before the start script.  This is true not only for start, but
for any script's name.  It is also possible to do a "postX" where X is
the name of another script.