Transpile build scripts using babel
Now that babel is installed, we can use it to transpile our build
scripts as well as our code so that everything all our code can leverage
the power of the new JavaScript features.

This simply requires changing calls to 'node' to instead call
'babel-note' when the scripts are invoked.