Add source maps and a breakpoint
Source maps map our bundled and/or minified source files back to the
original development files to enable debugging.

It's worth noting that the source maps are only generated when the
devtools are opened.  This can even be done in a production environment,
but I'm not sure how that would work with end users.

I also added a brekpoint to the source by simply adding a line with the
text 'debugger;', which is AWESOME!  Chrome automatically stopped at my
breakpoint and pulled up the original file via the source map!