Add npm scripts for ES Lint
I have also explicitly disabled ES Lint from enforcing the no-console
rule in my srcServer.js file, because I'm ok with printing to the
console in a development server.

I added a single line disable to startMessage.js, because
that's kind of the purpose of this dumb file.

I additionally added an ES Lint watch script so that ES Lint re-runs any
time any files change.  This has also been added to the start script, so
that ES Lint starts up when everything else does and is always running.