Add 'real' mocha test
It's worth noting that the second parameter to the it() function can be
a function with no parameters.  This is what you would do for a
synchronous test.

Because this test is asyncronous, because we are passing a callback to
jsdom.env, it is necessary to invoke the done() method to tell mocha
that our test is complete and it is now allowed to evaluate our expect.

The test will magically pass if you remove the done parameter and change
the .equal(string) string to something else.