Make calls to a mock API to fill HTML table
The API is currently just an endpoint on our development server,
returning static JSON.

With this change, I also eliminated the js/ and css/ folders that I had
created.  Similarily to the reasoning to put the tests right next to the
source files, I think that it's benefitial to have the CSS files right
next to the JS files.  This allows for simple import paths,
non-duplication of directory structures, and makes it obvious that the
files all go together.