Allow deletion of a user
It's a little confusing what is actually happening in the userApi right
now.  When fetch() is called, the base url that is returned points to
localhost:3001.  That is where we have json server running.

So when the delete link is clicked, a DELETE request is made to json
server with the path /users/id.  Because our mock schema has a top level
node of 'users', it realizes that we're trying to act on that object.
It must then somehow match the ID, and because it's a DELETE request,
that node is deleted.