Use specific query parameter to enable mock api
Instead of just checking for 'localhost', we now check for an explicity
query string parameter.  This allows us to easily switch between the
mock api (json-server) and the "real" api (express in srcServer.js).

Although there clearly is no "real" api at the moment, the express
server in srcServer.js is taking its place for the moment.