Split the production app into multiple bundles
Webpack refers to these bundles as 'chunks'

We've split the app into two bundles, one for our code and one for our
third party code.  This takes advantage of user side caching so that the
user does not have to re-download the third party code if only our code
has changed.

A large app could also be configured to be bundled by page.  This allows
the initial page to load faster, because the initial download only
consists of the code for the first page.