Implement cache-busting using MD5 hashes on filenames
Webservers can be configured to add headers to files to tell the browser
to not re-download them for up to a year. This is great because it saves
bandwidth and speeds page load times.  However, if that's done you need
a way of forcing the browser to redownload your updated code.

We do that here by adding deterministic hashes our filenames.  The
hashes are generated based on the contents of the file, thus the hash is
guaranteed to stay exactly the same until the contents of the file
changes because we've updated the code.  Neat!