Use hash routing instead of browser routing
Browser routing has an issue (cannot GET) if you try to go directly
to a URL without clicking on a link.  This is because all the react
routing is handles client side, and by going directly to a URL, that
request URL is given to the server that doesn't know how to handle it.

I tried to add a catch all "app.get("*", (req, res) => {}" to my server,
and that fixed the aforementioned problem, but it causes react router to
not render the path.  I just got a blank screen with only my

<div id="app"></div>

rendered on it.  Although I like BrowserHistory better for its cleaner
URLs (without the /#/) HashHistory fixed this somewhat major problem for
the time being so I'm going with it.

See here for more information on this:

    https://stackoverflow.com/a/36623117/2517147