Add CSS modules
A CSS Module is a CSS file in which all class names and animation names
are scoped locally by default.  This is done by adding hashes to the end
of every class name in the output CSS, thus making it unique throughout
your application, even though the original class name that you defined
can be repeated.

For more information, see here:
    https://github.com/css-modules/css-modules