Add SCSS support
Sass 3 introduces a new syntax known as SCSS which is fully compatible
with the syntax of CSS, while still supporting the full power of Sass.
This means that every valid CSS stylesheet is a valid SCSS file with the
same meaning.

Sass is a CSS syntax which is not as similar to classic CSS.  I prefer
SCSS because it still has the feel of CSS, but the power of SASS.

For more information on SCSS, see here:

    http://sass-lang.com/documentation/file.SCSS_FOR_SASS_USERS.html