Add react-bootstrap
This is a re-implementation of the imfamous bootstrap library for react.
Importing bootstrap into a CSS modules, react project in a usable way is
far more difficult than it should be.  This project removes the need for
that and also adds a lot of other functionality.

Note that I also added a package called react-router-bootstrap.  This
allows me to wrap react-bootstrap components in a component similar to
the NavLink from react-router-dom, such that navigation still works as
its supposed to, without having to hardcode routes.