Add manage course page
Note that I disabled the rule 'react/jsx-nobind' from ES Lint.  I did
that because the simplest way to use react router v4 to pass props to a
component is to call the render function of the Route instead of the
Component function, and pass the rendered component directly in 'html'.

See the CoursesPage for an example of this.