Organize npm dependencies
A "dependency" is anything that is required to build or run the
application.  A "devdependency" is anything else.  By putting everything
required to build or run the app in "dependencies", it is possible to
successfully run the app with "npm install --prod", which will be
slightly faster by not having to install those packages that are purely
for development/testing.