Allow importing of SCSS modules in typescript
This article really saved the day on the subject:
https://medium.com/@sapegin/css-modules-with-typescript-and-webpack-6b221ebe5f10

In short, the easy way is to use require instead of the ES6 import.
This still workes fine with webpack, but causes the typescript compiler
to basically leave it alone.

I of course hate that though, and want to use the import syntax.  To get
around that I use typings-for-css-modules-loader instead of css-loader,
which on the fly generates .scss.d.ts files to make typescript happy.