Allow navigation directly to a course
Previously, if you went to a specific course and then refreshed the
page, it would be blank.  That would happen because the AJAX call to get
the courses haden't populated yet, so the prop passed to the
ManageCoursePage would be null.

This is addressed by implementing the componentWillReceiveProps method
in the ManageCoursePage to change the state of the component when that
AJAX call comes through.