Setup architecture for react tests
Here we're using ts-mocha to allow us to wrote tests in the same
language that we're writing the rest of the app in.

In order to test react, we're using enzyme, which uses ReactTestUtils
under the covers.  Enzyme is an abstraction of ReactTestUtils, providing
a much friendlier api.

Mocha allows you to choose your own assertion library, and I have chai.
Chai then allows you to choose an assertion syntax, of which I choose
the 'should' syntax.  'Should' adds a should() method to the
Object.prototype to allow for a very natural syntax.