Determine which API to use based on webpack config
I installed a useful package called ifdef-loader.  This is a loader for
webpack that allows for directives in the code that will only include
certain lines based on conditions defined in the loader declaration in
the webpack config.

This allows for us to make requests to the mock API (on port 3001) for
debug, and the "real" API (on the same port as the app is deployed from)
for production, without some annoying query parameter that doesn't
really work.