fix: Typography in the application
The goal of this commit was to accomplish two things:

  1. Have the entire application use REM units for typography instead of
  px units.

  2. Have the entire application's font size respond to changes in the
  browsers default font size.

REM is a relative unit, based on the font size set in the html tag of
the page.  When REM units are used, the font sizes throughout the
application are then relative to the size set in the root html tag.
All one then has to do is make the entire application respond to changes
in the browser's default font size is set the font size of the root html
tag to a percentage.  Setting it to 100% would make 1rem equal to the
browser's default font size.

The only trouble with that is it makes the calculation of rem -> px
difficult when designing for the typical 16px default browser font size.
To deal with that, instead of setting the root html tag's font size to
100%, I set it to 10/16 as a percentage.  This makes these calculations
easy, where 1.6rem will be 16px on a browser with the default font size
of 16px.  But because the value is a percentage, it still
increases/decreases as the browser's font size changes.