feat: Add notifications that can be created from a reducer
In a perfect world, the React components themselves would not have to
ever show notifications.  Notifications are a result of some action
taking place in the application.  Because generic "actions" will almost
always map 1:1 with Redux actions, it would make sense to be able to
kick off notifications from Redux reducers.

This is now possible with notistack notifications.

It works by having a React component called <SnackbarNotifier /> exist
inside App.tsx.  This component checks for changes to a new
notifications array in the Redux state.  When it finds a new
notification, it tells notistack to show it, and then dispathes a Redux
action to remove the notification from the Redux state.  The
src/redux/notifications/reducer.ts then takes care of carrying out that
action.

The notification is removed almost immediately from the Redux state, but
that doesn't at all affect how long it's on the screen for, because this
is entirely determined by how notistack is configured once its
enqueueSnackbar() function is called.

---

Note that this can be combined with the createThunk() function (from
reduxUtil.ts) to show notifications when an error happens talking to the
api.  Simply throw an error in the api method, and createThunk will
catch this and dispatch the appropriate failure action with the error
message as the only parameter.

If the fail action is designed to take one parameter, that parameter can
then be accessed in the reducer and a notification created in the Redux
state that will be shown thanks to the <SnackbarNotifier /> (as
described above).