refactor: Make dev builds WAY faster by removing ts compiler
The awesome-typescript-loader is what slows down my builds to much.  I
removed it and replaced it with the @babel/typescript babel preset.
This preset just removes the extra type information from the ts files,
turning them into JS files.

Note that this means that it no longer does type checking.  I'm relying
on my editor to primarily take care of this.  But I also added an npm
run script 'yarn check-types' to run the typescript compiler to do type
checking.  Also, the production build still uses
awesome-typescript-loader for type checking, because speed isn't much of
an issue for that build.

I primarily followed this blog post for guidelines on how to perform
this process:

  https://iamturns.com/typescript-babel/